Add websocket support for KVM streaming
This was quite a pain to figure out.

Go's http.Request.Write() uses textproto.CanonicalMIMEHeaderKey()
which canonicalizes "Sec-WebSocket-*" headers to "Sec-Websocket-*"
(lowercase 's' in socket). These headers are supposed to be case
insensitive per RFC 7230 (the HTTP 1.1 RFC), but the server I'm
working with doesn't abide this: it'll respond to the request with a
503 response.

Instead, it needs WebSocket headers to be spelled case-sensitively
identical to how they're seen in the WebSocket RFC 6455:
"Sec-WebSocket-*" will not do, it must be `Sec-WebSocket-*`.

To avoid Go's header canonicalization, I see no other choice but work
the request bytes via something like `fmt`.