Get around blocking on touch stalling all requests
Previously, the locking was coarse enough that if any bastion needed a
new connection, blocking on touch, it would pause traffic being sent
to *any* bastion.

So, carefully manipulate locks here to unlock before doing the slow
step of getting material signed by the ssh-agent, which is responsible
for blocking on touch-verification.