Add response rewriting to get around hard-coded TLS protocol strings
At least one version of the BMC we interface with will always request
`wss://` in its javascript, even if the page is running via `http`
(since `mews` terminates TLS).  This will bounce off the `mews` proxy
with a connection reset, since `mews` is not expecting a TLS
connection to its HTTP server.

So, try this expedient but questionable hack: dumbly go through the
source code of html and javascript and rewrite references to `wss` to
`ws` and even `https` to `http` if it is to the current hostname.  It
works for the BMC software I care about right now: I can load their
HTML5 video pages.

But, it's a rather "practical" solution.  What if it becomes unwieldy?

In principle, `mews` could listen on TLS but that raises the problem
of adding certificates to the browser trust chain, whereas the whole
idea of `mews`'s configuration YAML can pin all the certificates and
act as a neat, and very exact form of trust chain.  Unfortunately it's
not possible to tell a browser "trust this CA only for *.localhost"
either.  But, if it comes to that, at least each client computer could
trust one, completely local CA that only will autosign `foo.localhost`
hosts.  I don't love it, but there's a way.