Shuffle bastion list only on config load
This misplacement of code defeats the intended goal, which is a
randomized round-robin between bastion servers to reach an upstream.
Instead it is closer to choosing a random server every time, which is
what this code was intended to prevent.

By shuffling just once and incrementing the currentIndex bounded by a
modulus, we get the intended behavior.