Add advanced SSH agent discovery
Previously, `mews` checked only the environment variable
`SSH_AUTH_SOCK`. This variable is not always used: in particular,
Secretive on macOS guides you to use `IdentityAgent` in `.ssh/config`.

OpenSSH has `.ssh/config` take precedence over `SSH_AUTH_SOCK`, so
adding an analogous feature to `mews` without going further would
deprive the user of a way to manipulate the agent.  We have one user
that relies on this.

Thus, add a flag as well.  This completes a tower of precedence that
mimics OpenSSH's:

1. Environment Variable: `SSH_AUTH_SOCK`
2. Configuration file: `.ssh/config`
3. Flag: `-agent=<SOMEPATH>`

Where the flag has the highest precedence.